requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
SQLAlchemy==2.0.34
pyserial==3.5
pytest==8.3.2
Faker==29.0.0
